lsp: panic on invalid offsets
Having mapper methods not return Options is a lot more ergonomic. In theory, neither our code nor the language server protocol should generate invalid offsets, so it is correct to panic in this case. It may still be desirable instead generate an internal server error to protect against bugs, but we'll defer that to a future change.